+++ include/NinjaHSM/StateMachine.hpp
@@ -1,15 +1,18 @@
 #pragma once
 
-#include <array>
-#include <iostream>
+#include <cstdint>
 
 #include "State.hpp"
 
-
 namespace NinjaHSM {
 
-constexpr uint32_t MAX_RECURSION_COUNT = 10;
-constexpr uint32_t MAX_STATES = 100;
+/**
+ * The maximum number of times transitionTo() can be called recursively.
+ * This is to prevent infinite recursion in case of a bug. transitionTo()
+ * is called recursively if it called within a state's entry() or exit()
+ * methods.
+ */
+constexpr uint32_t MAX_RECURSION_COUNT = 50;
 
 template <typename Event>
 class StateMachine {
@@ -102,15 +105,9 @@ protected:
         transitionToCalled = true;
         maxRecursionCount++;
         if (maxRecursionCount > MAX_RECURSION_COUNT) {
-            std::cout << "Max recursion count reached. Aborting transition." << std::endl;
             return;
         }
         uint32_t ourRecursionCount = maxRecursionCount;
-        std::cout << "transitionTo() called. Our recursion count is: " << ourRecursionCount << std::endl;
-        if (currentState == nullptr) {
-            std::cout << "Current state is null." << std::endl;
-        }
-        std::cout << "Destination state: " << state->name << std::endl;
 
         // Rename just for readability below
         State<Event>* destinationState = state;
@@ -119,22 +116,18 @@ protected:
         // If the new destination state is a child of the previous entry() function,
         // we don't want to re-call the entry() function (we assume the state was entered).
         if (calledEntryState != nullptr && isChildOf(calledEntryState, destinationState)) {
-            std::cout << "Just called entry state is a parent of the destination state.Assuming state was entered." << std::endl;
             currentState = calledEntryState;
             calledEntryState = nullptr; // Clear flag
         }
 
         if (calledExitState != nullptr && !isChildOf(calledExitState, destinationState)) {
-            std::cout << "Just called exit state is not a child of the destination state. Assuming state was exited." << std::endl;
             currentState = calledExitState->parent;
             calledExitState = nullptr; // Clear flag
         }
 
         if (currentState == destinationState) {
-            std::cout << "Current state is the same as the destination state. Exiting and entering again." << std::endl;
             currentState->exit();
             if (ourRecursionCount != maxRecursionCount) {
-                std::cout << "Recursion detected. Aborting transition." << std::endl;
                 goto END;
             }
             currentState = currentState->parent;
@@ -142,12 +135,6 @@ protected:
 
         // This loop handles one entry or exit per iteration.
         while (currentState != destinationState) {
-            if (currentState == nullptr) {
-                std::cout << "Current state is null." << std::endl;
-            } else {
-                std::cout << "Current state: " << currentState->name << std::endl;
-            }
-            std::cout << "Destination state: " << destinationState->name << std::endl;
             // Logic:
             // Search for the current state in the tree containing the destination
             // state and all of it's parents. If the current state is found,
@@ -170,34 +157,28 @@ protected:
             if (foundCurrentStateInDestinationBranch) {
                 // We've found the current state in the destination branch.
                 // Move down one state.
-                std::cout << "Found current state in destination branch. Moving down one state to: " << stateInDestinationBranch->name << std::endl;
                 calledEntryState = stateInDestinationBranch;
                 stateInDestinationBranch->entry();
                 calledEntryState = nullptr;
                 if (ourRecursionCount != maxRecursionCount) {
-                    std::cout << "Recursion detected. Aborting transition." << std::endl;
                     break;
                 }
-                std::cout << "Setting current state to: " << stateInDestinationBranch->name << std::endl;
                 currentState = stateInDestinationBranch;
                 continue;
             }
 
             // If we get here, we need to exit the current state.
-            std::cout << "Need to exit current state." << std::endl;
             // Transition to the top most parent of the destination state.
             calledExitState = currentState;
             currentState->exit();
             calledExitState = nullptr; // Clear flag
             if (ourRecursionCount != maxRecursionCount) {
-                std::cout << "Recursion detected. Aborting transition." << std::endl;
                 break;
             }
             currentState = currentState->parent; // This might be nullptr
         }
 
         END:
-        std::cout << "transitionTo() finished. Our recursion index is: " << ourRecursionCount << ", current state is: " << currentState->name << std::endl;
 
         // If we are at the top of the recursion, reset the recursion index so it's
         // ready for the next non-recursive transitionTo() call.